Histogram equalization and matching in DIP
In order to perform Histogram equalization of an image we need to calculate the PDF (Probability Density Functions) and then CDF (Cumulative Distribution Function) and Finally calculate the Histogram equalization. And for matching, we do the the safe procedure (PDF and then CDF) of a reference image and do the matching with the equalized image.